moved encrypt() & decrypt() to Chat.java
